+++ src/cc/include/swcdb/core/comm/HeaderBufferInfo.h
@@ -1,4 +1,4 @@
-/* 
+/*
  * SWC-DB© Copyright since 2019 Alex Kashirin <kashirin.alex@gmail.com>
  * License details at <https://github.com/kashirin-alex/swc-db/#license>
  */
@@ -17,26 +17,26 @@ namespace SWC { namespace Comm {
 struct BufferInfo final {
 
   BufferInfo();
-    
+
   void reset();
 
-  size_t encoded_length() const;
+  uint8_t encoded_length() const;
 
   void encode(uint8_t** bufp) const;
 
   void decode(const uint8_t** bufp, size_t* remainp);
 
   void encode(Core::Encoder::Type encoder, StaticBuffer& data);
-    
+
   void decode(int& err, StaticBuffer& data) const;
 
   void print(std::ostream& out) const;
-  
+
   uint32_t            size;        //!< Buffer size
   Core::Encoder::Type encoder;     //!< Buffer Encoder
   uint32_t            size_plain;  //!< Buffer set if Encoder not PLAIN
   uint32_t            chksum;      //!< Buffer checksum
-  
+
 } __attribute__((packed));
 
 
@@ -46,6 +46,6 @@ struct BufferInfo final {
 
 #ifdef SWC_IMPL_SOURCE
 #include "swcdb/core/comm/HeaderBufferInfo.cc"
-#endif 
+#endif
 
 #endif // swcdb_core_comm_HeaderBufferInfo_h
